test: split new stdlib-reachability tests into a sibling file
CI failure: the previous commit's additions pushed
tests/test_contract_evaluation.py to 2008 lines, over the
AI-readiness gate's 2000-line hard cap. Moved
TestPublicModeDirectlyReferencedStdlib into a new sibling file,
following the existing test_contract_evaluation_exports.py /
test_contract_evaluation_not_applicable.py split pattern for the
same file.

The check_tier_accuracy.py "ModuleNotFoundError: No module named
'elftools'" failure in the same CI run was a pure cascade from this:
the ai-readiness job's "Install package" step (pip install -e
".[dev]") has no `if: always()`, so once the file-size ERROR failed
the prior step, the install step was skipped -- starving the later
`if: always()` check_tier_accuracy.py step of its dependency. No
separate fix needed; confirmed by reading ci.yml's step ordering.

Verified: both files pass locally, ai-readiness gate reports 0
errors (was 1), ruff/mypy clean.
