+++ tests/test_contract_evaluation.py
@@ -1016,112 +1016,6 @@ class TestPublicModeConservativeRetentionIsNotConfirmation:
         assert decision.reason_code == "required_evidence_incomplete"
 
 
-class TestPublicModeDirectlyReferencedStdlib:
-    """``directly_referenced_stdlib_old``/``directly_referenced_stdlib_new`` --
-    a second confirmation source alongside ``public_types``, for a stdlib
-    type a public signature names outright (``public_types`` deliberately
-    excludes stdlib types as non-ABI-surface toolchain internals)."""
-
-    def test_finding_on_a_directly_referenced_stdlib_type_is_confirmed(
-        self,
-    ) -> None:
-        # public_types has nothing to say about a stdlib type at all
-        # (surface.py never walks into it) -- confirmation must come purely
-        # from the reachability evidence.
-        snap = AbiSnapshot(library="l", version="1", functions=[_fn("api")])
-        s = compute_public_surface(snap)
-        c = Change(
-            kind=ChangeKind.TYPE_SIZE_CHANGED,
-            symbol="vector<int, std::allocator<int> >",
-            description="",
-        )
-        decision = evaluate_change_contract_relevance(
-            c,
-            s,
-            s,
-            mode=ContractMode.PUBLIC,
-            directly_referenced_stdlib_old=frozenset(
-                {"vector<int, std::allocator<int> >"}
-            ),
-            directly_referenced_stdlib_new=frozenset(
-                {"vector<int, std::allocator<int> >"}
-            ),
-        )
-        assert decision.relevance is ContractRelevance.IN_CONTRACT
-        assert decision.reason_code == "public_root_membership"
-
-    def test_a_nested_directly_referenced_type_does_not_confirm_its_container(
-        self,
-    ) -> None:
-        # Regression (Codex review, fresh evidence): matching was originally
-        # containment (does the referenced identity appear *inside* the
-        # finding's own spelling), which let an unrelated, independently
-        # directly-referenced std::allocator<int> confirm a TYPE_SIZE_CHANGED
-        # finding on the unrelated, larger
-        # std::vector<int, std::allocator<int> > purely because one spelling
-        # embeds the other -- reproduced empirically before the fix (exact
-        # matching only) landed. The vector itself is genuinely NOT among
-        # the directly-referenced identities here.
-        snap = AbiSnapshot(library="l", version="1", functions=[_fn("api")])
-        s = compute_public_surface(snap)
-        c = Change(
-            kind=ChangeKind.TYPE_SIZE_CHANGED,
-            symbol="vector<int, std::allocator<int> >",
-            description="",
-        )
-        decision = evaluate_change_contract_relevance(
-            c,
-            s,
-            s,
-            mode=ContractMode.PUBLIC,
-            directly_referenced_stdlib_old=frozenset({"std::allocator<int>"}),
-            directly_referenced_stdlib_new=frozenset({"std::allocator<int>"}),
-        )
-        assert decision.relevance is ContractRelevance.UNKNOWN_UNRESOLVED
-        assert decision.reason_code == "required_evidence_incomplete"
-
-    def test_new_side_only_evidence_does_not_confirm_an_old_side_authoritative_finding(
-        self,
-    ) -> None:
-        # ADR-049 D4: a modification (not in ADDITION_KINDS) is judged by the
-        # old side alone -- new-side-only reachability evidence must not
-        # manufacture confidence about an old, unresolved obligation, same
-        # as the existing public_types side-authority rule.
-        snap = AbiSnapshot(library="l", version="1", functions=[_fn("api")])
-        s = compute_public_surface(snap)
-        c = Change(
-            kind=ChangeKind.TYPE_SIZE_CHANGED,
-            symbol="vector<int, std::allocator<int> >",
-            description="",
-        )
-        decision = evaluate_change_contract_relevance(
-            c,
-            s,
-            s,
-            mode=ContractMode.PUBLIC,
-            directly_referenced_stdlib_old=frozenset(),
-            directly_referenced_stdlib_new=frozenset(
-                {"vector<int, std::allocator<int> >"}
-            ),
-        )
-        assert decision.relevance is ContractRelevance.UNKNOWN_UNRESOLVED
-        assert decision.reason_code == "required_evidence_incomplete"
-
-    def test_omitting_the_params_is_unaffected(self) -> None:
-        # Default None on both -- every pre-existing caller's behaviour is
-        # unchanged.
-        snap = AbiSnapshot(library="l", version="1", functions=[_fn("api")])
-        s = compute_public_surface(snap)
-        c = Change(
-            kind=ChangeKind.TYPE_SIZE_CHANGED,
-            symbol="vector<int, std::allocator<int> >",
-            description="",
-        )
-        decision = evaluate_change_contract_relevance(c, s, s, mode=ContractMode.PUBLIC)
-        assert decision.relevance is ContractRelevance.UNKNOWN_UNRESOLVED
-        assert decision.reason_code == "required_evidence_incomplete"
-
-
 class TestPublicModeMemberLevelConfirmation:
     """A member-level finding must be confirmed against its *owner* type.
